requirements.txt
catboost==1.2.7
coverage==7.6.12
dvc==3.59.2
flask==3.1.0
matplotlib==3.10.0
numpy==1.26.4
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2